fix: merge hooks folders into on
