+++ packages/chap-lib/i18n/en.pot
@@ -5,19 +5,17 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2025-04-22T11:00:23.331Z\n"
-"PO-Revision-Date: 2025-04-22T11:00:23.332Z\n"
+"POT-Creation-Date: 2025-04-24T13:47:59.394Z\n"
+"PO-Revision-Date: 2025-04-24T13:47:59.395Z\n"
 
-msgid ""
-"Prediction for ${findOrgUnitName(\n"
-"                                    ou,\n"
-"                                    dataValues\n"
-"                                )}"
-msgstr ""
-"Prediction for ${findOrgUnitName(\n"
-"                                    ou,\n"
-"                                    dataValues\n"
-"                                )}"
+msgid "Median Prediction for {{predictionTargetName}}"
+msgstr "Median Prediction for {{predictionTargetName}}"
+
+msgid "Prediction for {{orgUnitName}}"
+msgstr "Prediction for {{orgUnitName}}"
+
+msgid "Prediction for {{predictionTargetName}} for {{orgUnitName}}"
+msgstr "Prediction for {{predictionTargetName}} for {{orgUnitName}}"
 
 msgid "Quantile median"
 msgstr "Quantile median"
