+++ packages/prediction-app/src/features/job-results/JobResultPanel/JobResultPanelItem/JobResultPanelItem.tsx
@@ -12,7 +12,7 @@ import {
     IconMore16,
     Menu,
     MenuItem,
-    Popper
+    Popper,
 } from '@dhis2/ui'
 import { JobResult } from '../../interfaces/JobResult'
 import { useClickOutside } from '../../../../hooks/useClickOutside'
@@ -26,7 +26,7 @@ interface JobResultPanelItem {
     onClickRemove: any
 }
 
-const JobResultPanelItem = ({ 
+const JobResultPanelItem = ({
     i,
     jobResult: jobResult,
     onClickEvaluateDataset,
@@ -34,12 +34,11 @@ const JobResultPanelItem = ({
     onClickImport,
     onClickViewLogs,
     onClickRemove,
-  }: JobResultPanelItem) => {
-
+}: JobResultPanelItem) => {
     const [showMenu, setShowMenu] = useState(false)
     const buttonRef = useRef<HTMLDivElement>(null)
     const menuRef = useRef<HTMLDivElement>(null)
-    
+
     // Close menu when clicking outside menu or button
     useClickOutside(menuRef, (event) => {
         // Only close if clicked outside the menu and not on the trigger
@@ -86,20 +85,14 @@ const JobResultPanelItem = ({
     return (
         <div key={i} className={styles.boxPanel}>
             <div className={styles.jobResultPanelContainer}>
-                <div className={styles.flexMedium}>
-                    {jobResult.name}
-                </div>
+                <div className={styles.flexMedium}>{jobResult.name}</div>
                 <div className={styles.flexMedium}>
                     {jobResult.created.toDateString() +
                         ', ' +
-                        jobResult.created
-                            .toLocaleTimeString()
-                            .slice(0, 5)}
+                        jobResult.created.toLocaleTimeString().slice(0, 5)}
                 </div>
                 <div className={styles.flexMedium}>
-                    <span
-                        className={getStatusColor(jobResult.status)}
-                    >
+                    <span className={getStatusColor(jobResult.status)}>
                         {getStatusText(jobResult.status)}
                     </span>
                 </div>
@@ -108,14 +101,11 @@ const JobResultPanelItem = ({
                         {
                             prediction: (
                                 <>
-                                    {jobResult.status == 
-                                    'SUCCESS' && (
+                                    {jobResult.status == 'SUCCESS' && (
                                         <Button
                                             icon={<IconLaunch24 />}
                                             onClick={() =>
-                                                onClickImport(
-                                                    jobResult.result
-                                                )
+                                                onClickImport(jobResult.result)
                                             }
                                             small
                                         >
@@ -126,12 +116,13 @@ const JobResultPanelItem = ({
                             ),
                             dataset: (
                                 <>
-                                    {jobResult.status == 
-                                    'SUCCESS' && (
+                                    {jobResult.status == 'SUCCESS' && (
                                         <Button
                                             icon={<IconRuler24 />}
-                                            onClick={() => 
-                                                onClickEvaluateDataset(jobResult.result)
+                                            onClick={() =>
+                                                onClickEvaluateDataset(
+                                                    jobResult.result
+                                                )
                                             }
                                             small
                                         >
@@ -142,12 +133,15 @@ const JobResultPanelItem = ({
                             ),
                             evaluation: (
                                 <>
-                                    {jobResult.status == 
-                                    'SUCCESS' && (
+                                    {jobResult.status == 'SUCCESS' && (
                                         <Button
-                                            icon={<IconVisualizationAreaStacked24 />}
-                                            onClick={() => 
-                                                onClickViewEvaluation(jobResult.result)
+                                            icon={
+                                                <IconVisualizationAreaStacked24 />
+                                            }
+                                            onClick={() =>
+                                                onClickViewEvaluation(
+                                                    jobResult.result
+                                                )
                                             }
                                             small
                                         >
@@ -159,23 +153,22 @@ const JobResultPanelItem = ({
                         }[jobResult.type]
                     }
 
-                    {jobResult.status != 'SUCCESS' 
-                        && (
-                            <Button
-                                icon={<IconArchive24 />}
-                                onClick={() => 
-                                    onClickViewLogs(jobResult.id)
-                                }
-                                small
-                            >
-                                Logs
-                            </Button>
+                    {jobResult.status != 'SUCCESS' && (
+                        <Button
+                            icon={<IconArchive24 />}
+                            onClick={() => onClickViewLogs(jobResult.id)}
+                            small
+                        >
+                            Logs
+                        </Button>
                     )}
 
                     <div ref={buttonRef}>
                         <ConditionalTooltip
                             enabled={jobResult.status === 'STARTED'}
-                            content={i18n.t('Job is still running. Please wait until it is finished before doing any actions.')}
+                            content={i18n.t(
+                                'Job is still running. Please wait until it is finished before doing any actions.'
+                            )}
                         >
                             <Button
                                 small
@@ -190,7 +183,9 @@ const JobResultPanelItem = ({
                         <Popper reference={buttonRef} placement="top-end">
                             <div ref={menuRef} className={styles.popupMenuDiv}>
                                 <Menu>
-                                    {['SUCCESS', 'FAILURE'].includes(jobResult.status) && (
+                                    {['SUCCESS', 'FAILURE'].includes(
+                                        jobResult.status
+                                    ) && (
                                         <MenuItem
                                             label="Delete"
                                             icon={<IconDelete24 />}
@@ -205,7 +200,6 @@ const JobResultPanelItem = ({
                             </div>
                         </Popper>
                     )}
-
                 </div>
             </div>
         </div>
