+++ packages/prediction-app/src/features/send-chap-data/components/DownloadData/DownloadAnalyticsData.tsx
@@ -1,8 +1,8 @@
 import { saveAs } from 'file-saver'
 import JSZip from 'jszip'
 import React, { Dispatch, SetStateAction, useEffect, useState } from 'react'
-import useAnalyticRequest from '../../../../new-view-hooks/useAnalyticRequest'
-import useGeoJson from '../../../../new-view-hooks/useGeoJson'
+import useAnalyticRequest from '../../../../hooks/useAnalyticRequest'
+import useGeoJson from '../../../../hooks/useGeoJson'
 import { ModelFeatureDataElementMap } from '../../../../interfaces/ModelFeatureDataElement'
 import styles from './DownloadAnalyticsData.module.css'
 import { ErrorResponse } from '../../interfaces/ErrorResponse'
