+++ packages/prediction-app/src/hooks/useGetRoute.tsx
@@ -1,25 +1,27 @@
-import { useState } from 'react'
-import { useDataQuery } from '@dhis2/app-runtime'
+import { useState } from "react";
+import { useDataQuery } from "@dhis2/app-runtime";
 
 const REQUEST = {
-    routes: {
-        resource: 'routes',
-        params: {
-            paging: false,
-            filter: `code:eq:chap`,
-            fields: '*',
-        },
+  routes : {
+    resource: "routes",
+    params: {
+      paging : false,
+      filter : `code:eq:chap`,
+      fields : "*"
     },
+  }
 }
 
 const useGetRoute = () => {
-    const { data: route, loading, error } = useDataQuery(REQUEST)
 
-    return {
-        route: (route as any)?.routes?.routes[0],
-        error,
-        loading,
-    }
-}
+  const { data : route, loading, error } = useDataQuery(REQUEST);
 
-export default useGetRoute
+
+  return {
+    route : (route as any)?.routes?.routes[0],
+    error,
+    loading,
+  };
+};
+
+export default useGetRoute;
