+++ packages/prediction-app/src/hooks/useOrgUnits.tsx
@@ -1,23 +1,28 @@
-import { useState } from 'react'
-import { useDataQuery } from '@dhis2/app-runtime'
+import { useState } from "react";
+import { useDataQuery } from "@dhis2/app-runtime";
 
 const REQUEST = {
-    orgUnits: {
-        resource: 'organisationUnits',
-        params: {
-            paging: false,
-            fields: ['id', 'displayName'],
-        },
+ orgUnits : {
+    resource: "organisationUnits",
+    params: {
+      paging : false,
+      fields : ['id','displayName']
     },
+  }
 }
 
 const useOrgUnits = () => {
-    const { loading, error, data } = useDataQuery(REQUEST)
-    return {
-        orgUnits: (data as any)?.orgUnits,
-        error,
-        loading,
-    }
-}
+  const [orgUnits, setOrgunits] = useState<{organisationUnits : any}>();
 
-export default useOrgUnits
+  const { loading, error } = useDataQuery(REQUEST, {
+    onComplete: (data : any) => setOrgunits(data.orgUnits),
+  });
+
+  return {
+    orgUnits,
+    error,
+    loading,
+  };
+};
+
+export default useOrgUnits;
